Add full server functionality.  Passes all autobahn tests (v0.5)
BIG change.  Rename 'network' module to 'protocol' because that's a better description.  Add entire new 'server' module that combines 'event' and 'protocol' modules.  Add new 'inlist.h' for in-place linked lists.  Fixes a nasty nasty bug that lead to buffer overflows in darray.  Add an 'echoserver' which is for running the autobahntests against.