server: Add should_stop callback
Add the ability for the user of the server module to specify a
should_stop callback. It will be called periodically, and if it returns
true, the server will do the same thing as server_stop.

This is useful if you know you want to stop the server but don't have
access to the server object itself to call server_stop (that is, a
signal handler).