protocol: zero copy write
This change implements 'zero copy write' optimizations for all protocol
functions that currently add to the write buffer. You can pass in an fd,
which will be written to directly if there are no bytes on the write
buffer. In autobahn performance tests, there appears to be no real
performance gain. I suspect this is because the extra memcpy is such a
small part of most operations, and heelhook is massivly bound by
read/write syscall overhead. Putting this on another branch for further
investigation.