protocol: Fix bug in utf8 validation
It was possible for a perfectly fine utf8 message to be passed as valid,
or visa versa, due to prematurely resetting validator state. To
reproduce the issue, the following sequence of messages had to be read off the
read_buffer:

1. TEXT MESSAGE FRAGMENT 1
2. PING OR PONG MESSAGE
3. TEXT MESSAGE FRAGMENT 2

This is a perfectly valid sequence of frames according to the websocket
standard. In protocol_read_msg, the parser state was being reset after
the ping or pong message in (2). The fix is to only reset validator
state when a text message has been completed.