hhlog: Fix improper use of snprintf
When there is not enough space in the buffer to print all characters,
snprintf will print out the number of characters that WOULD have been
written if there was space. hhlog didn't properly take this into
account, so instead of truncating the printed strings, a buffer overflow
could occur. This change fixes that, and ups the line length limit to
4K.