endpoint: More properly clean up memory
The memory reclamation logic around the read buffer was flawed if the
server received many ping or pong messages in a row, without a 'data'
message in between. The memory wouldn't be reclaimed until a 'data' message was
received. This change fixes that by keeping track of all memory processed during
the main parsing loop, and removes it afterwards.