server: Fix crash in server_conn_close
If server_conn_close was called on a connection that was currently in
the HANDSHAKE phase of the connection, we'd try to add an event callback
for an fd of -1, due to the results of this callstack:

    server_on_close
    deactivate_conn (endpoint.c)
    endpoint_close
    server_conn_close

After this, conn->fd would be -1 due to server_on_close.
server_conn_close would try to call queue_write which would in turn try
to add an event call backfor conn->fd.